cost-tracking: testing infrastructure + some fixes
This patch adds the infrastructure necessary to test the runtime's cost
tracking, including all necessary fixes to the underlying code allowing
this to be done.

Included are:

- Fixed costs dependant on the runtime length of lists, buffers, etc...
  to use the actual runtime length
- Fixed cost tracking table multiplexing based on clarity version
- Added testing for every word, with the notable exceptions of `contract-call`
  and `contract-of`, to be added in a future changeset
- Small, clear, fixes to the implementations of `StxTransferMemo` and
  `Sqrti`
- Made available cost points infinite by default
- Removal of parity check with the interpreter, as discussed

The following known defects still remain, to be addressed in future
changes:

- Use of the function introduced in #713 in the cost tracking code for
  arbitrarily sized values
- Fix map implementation to use the above for both keys and values

Both changes to be done in a future changeset to avoid making this
change (even) larger.

Resolves: #616
Resolves: #636